move config to fix implicit dependency on en
when creating a new `Chrono` instance, it uses the en config by default when none is provided in the constructor

importing this from `locales/en` creates an implicit dependency which breaks when using locales individually

moving this to a separate file eliminates this implicit dependency and prevents breakage
